ngsi-ld/data-models/openapi-schemas/ietf-interfaces-tester/pyproject.toml
[tool.poetry]
name = "ietf-interfaces-tester"
version = "0.1.0"
description = ""
authors = ["GIROS-DIT UPM <gi.giros@upm.es>"]

[tool.poetry.dependencies]
python = "^3.9.5"
ngsi-ld-models = ">=1.0.0"
fastapi = {extras = ["all"], version = "^0.95.2"}
ngsi_ld_client = { git = "https://github.com/giros-dit/python-ngsi-ld-client.git", branch = "1.6.1" }
pyyaml = "^6.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = ">=4.0.1"
ngsi-ld-models = { path = "../ngsi-ld-models", develop = true }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
